requirements_rl.txt
# ===================================================================
# File: requirements_rl.txt
# Additional requirements for RL experiments  
# ===================================================================

# RL Libraries
stable-baselines3==2.0.0
gymnasium==0.29.1
sb3-contrib==2.0.0

# Visualization and Analysis
tensorboard==2.15.1
wandb==0.16.1
seaborn==0.12.2
plotly==5.17.0

# Existing requirements (add if not already present)
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
matplotlib>=3.7.0
pyyaml>=6.0
opencv-python>=4.8.0
